gene_summary_22Q1

The `gene_summary` dataset contains the 22Q1 gene essentiality probabilities for select genes. This dataset can be loaded into the R environment with the `depmap_gene_summary` function.

gene_summary

Format

A data frame with 69746 rows (cell lines) and 7 variables:

entrez_id

Entrez ID# (e.g. 100316904)

gene_name

HUGO symbol (e.g. "SAP25")

dataset

which dataset this probability derives

dependent_cell_lines

number of dependent cell lines

cell_lines_with_data

number of cell lines with relevant dependency data

strongly_selective

Gene knockout is selective (not pan-lethal)

common_essential

common essential gene dependency

Source

DepMap, Broad Institute: https://depmap.org/portal/download/

Details

This data represents the `Gene Dependency Profile Summary.csv` file taken from the 22Q1 [Broad Institute](https://depmap.org/portal/api/download/gene_dep_summary) release.

Change log

- 22Q1: Initial dataset

- 22Q2: no change, no further releases are scheduled at this time.

References

Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A., Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer dependency map. Cell, 170(3), 564-576.

James M. McFarland, Zandra V. Ho, Guillaume Kugener, Joshua M. Dempster, Phillip G. Montgomery, Jordan G. Bryan, John M. Krill-Burger, Thomas M. Green, Francisca Vazquez, Jesse S. Boehm, Todd R. Golub, William C. Hahn, David E. Root, Aviad Tsherniak. (2018). Improved estimation of cancer dependencies from large-scale RNAi screens using model-based normalization and data integration. Nature Communications 9, 1.

Examples

if (FALSE) {
depmap_gene_summary()
}